fix: settings command overwrites entire JSON file
